requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.10.3
python-dotenv==1.0.1
twilio==9.3.7
PyJWT==2.9.0